Prevent duplicates in UltiSnipsEdit! due to g:UltiSnipsSnippetStorageDirectoryForUltiSnipsEdit
If there's a directory on the runtimepath that is a symlink to
g:UltiSnipsSnippetStorageDirectoryForUltiSnipsEdit (or the other way
round), the same snippet files will end up on the UltiSnipsEdit! list
twice, under two different paths. This commit prevents that, favoring
the path under g:UltiSnipsSnippetStorageDirectoryForUltiSnipsEdit.

Note that this does nothing to prevent potential duplicate snippet files
that arise from having duplicate symlinked entries on the runtimepath.
Those should be avoided by fixing your runtimepath so that it doesn't
have duplicate symlinked entries.